+++ segmenter_test.go
@@ -78,54 +78,9 @@ func TestSegment(t *testing.T) {
 	tt.Expect(t, "24", segments[3].end)
 }
 
-func TestSegmentS(t *testing.T) {
-	var seg Segmenter
-	seg.LoadDict("zh,testdata/test_dict.txt")
-	// seg.LoadDict()
-
-	dict := seg.Dictionary()
-	tt.Expect(t, "16", dict.maxTokenLen)
-	tt.Expect(t, "5.3250728e+07", dict.totalFrequency)
-
-	tt.Expect(t, "587881", seg.Dict.NumTokens())
-	text1 := "纽约帝国大厦, 旧金山湾金门大桥"
-	segStr := "纽约/ns 帝国大厦/nr ,/x  /x 旧金山湾/ns 金门大桥/nz "
-
-	tt.Expect(t, "纽约/ns 帝国大厦/nr ,/x  /x 旧金山湾/ns 金门大桥/nz ", seg.String(text1))
-	tt.Expect(t, "[纽约 帝国大厦 ,   旧金山湾 金门大桥]", seg.Slice(text1))
-
-	tt.Expect(t,
-		"纽约/ns 帝国/n 大厦/n 帝国大厦/nr ,/x  /x 金山/nr 旧金山/ns 湾/zg 旧金山湾/ns 金门/n 大桥/ns 金门大桥/nz ",
-		seg.String(text1, true))
-	tt.Expect(t, "[纽约 帝国 大厦 帝国大厦 ,   金山 旧金山 湾 旧金山湾 金门 大桥 金门大桥]", seg.Slice(text1, true))
-
-	segments := seg.Segment([]byte(text1))
-	tt.Expect(t, segStr, ToString(segments))
-	tt.Expect(t, segStr, ToString(segments, false))
-
-	segs := seg.ModeSegment([]byte(text1), true)
-	tt.Expect(t, segStr, ToString(segs, false))
-
-	tt.Expect(t, "6", len(segments))
-	tt.Expect(t, "0", segments[0].start)
-	tt.Expect(t, "6", segments[0].end)
-	tt.Expect(t, "6", segments[1].start)
-	tt.Expect(t, "18", segments[1].end)
-
-	text2 := []byte("留给真爱你的人")
-	segments2 := seg.Segment(text2)
-	tt.Expect(t, "留给/v 真爱/nr 你/r 的/uj 人/n ", ToString(segments2, false))
-
-	tt.Expect(t, "5", len(segments2))
-	tt.Expect(t, "0", segments2[0].start)
-	tt.Expect(t, "6", segments2[0].end)
-	tt.Expect(t, "6", segments2[1].start)
-	tt.Expect(t, "12", segments2[1].end)
-}
-
 func TestSegmentJp(t *testing.T) {
 	var seg Segmenter
-	SkipLog = true
+	// SkipLog = true
 	seg.LoadDict("data/dict/jp/dict.txt")
 	segments := seg.Segment(testH)
 
@@ -140,87 +95,7 @@ func TestSegmentJp(t *testing.T) {
 	tt.Expect(t, "15", segments[0].end)
 }
 
-func TestDictPaths(t *testing.T) {
-	paths := DictPaths("./dictDir", "zh,jp")
-	tt.Expect(t, "2", len(paths))
-
-	tt.Expect(t, "dictDir/dict/dictionary.txt", paths[0])
-	tt.Expect(t, "dictDir/dict/jp/dict.txt", paths[1])
-
-	paths1 := DictPaths("./dictDir", "zh, jp")
-	tt.Expect(t, "2", len(paths))
-	tt.Equal(t, paths, paths1)
-}
-
-func TestSegmentDicts(t *testing.T) {
-	var seg Segmenter
-	// seg.LoadDict("zh,jp")
-	seg.LoadDict("./data/dict/dictionary.txt,./data/dict/jp/dict.txt")
-
-	text1 := []byte("旧金山湾金门大桥")
-	segments := seg.Segment(text1)
-	tt.Expect(t, "旧金山湾/ns 金门大桥/nz ", ToString(segments, false))
-
-	tt.Expect(t, "2", len(segments))
-	tt.Expect(t, "0", segments[0].start)
-	tt.Expect(t, "12", segments[0].end)
-	tt.Expect(t, "12", segments[1].start)
-	tt.Expect(t, "24", segments[1].end)
-
-	segments = seg.Segment(testH)
-	tt.Expect(t, "こんにちは/感動詞 世界/n ", ToString(segments, false))
-	tt.Expect(t, "2", len(segments))
-	tt.Expect(t, "こん/名詞 こんにちは/感動詞 世界/n ", ToString(segments, true))
-	tt.Expect(t, "2", len(segments))
-	tt.Expect(t, "0", segments[0].start)
-	tt.Expect(t, "15", segments[0].end)
-
-	tt.Expect(t, "0", segments[0].Start())
-	tt.Expect(t, "15", segments[0].End())
-
-	token := segments[0].Token()
-	tt.Expect(t, "こんにちは", token.Text())
-	tt.Expect(t, "5704", token.Frequency())
-	tt.Expect(t, "感動詞", token.Pos())
-
-	var tokenArr []*Token
-	for i := 0; i < len(segments); i++ {
-		tokenArr = append(tokenArr, segments[i].Token())
-	}
-	tt.Expect(t, "こんにちは 世界 ", printTokens(tokenArr, 2))
-
-	tseg := token.Segments()
-	tt.Expect(t, "0", tseg[0].Start())
-	tt.Expect(t, "6", tseg[0].End())
-}
-
-func TestLargeDictionary(t *testing.T) {
-	err := prodSeg.LoadDict("zh,testdata/test_dict2.txt")
-	tt.Nil(t, err)
-
-	text1 := []byte("世界人口")
-	text2 := []byte("山达尔星新星联邦共和国联邦政府")
-
-	tt.Expect(t, "世界/n 人口/n ", ToString(prodSeg.Segment(text1), false))
-
-	tt.Expect(t, "世界/n 人口/n ", ToString(prodSeg.internalSegment(text1, false),
-		false))
-
-	tt.Expect(t, "世界/n 人口/n ",
-		ToString(prodSeg.internalSegment(text1, true), false))
-
-	tt.Expect(t, "山达尔星新星联邦共和国/ns 联邦政府/nt ",
-		ToString(prodSeg.internalSegment(text2, true), false))
-
-	tt.Expect(t, "山达尔星新星联邦共和国联邦政府/nt ",
-		ToString(prodSeg.internalSegment(text2, false), false))
-
-	tt.Expect(t, "达尔/nrt 星/n 山达尔星/nz 新星/nz 联邦/n 共和/nz 国/n 共和国/ns 联邦共和国/nt 山达尔星新星联邦共和国/ns 联邦/n 政府/n 联邦政府/nt 山达尔星新星联邦共和国联邦政府/nt ",
-		ToString(prodSeg.Segment(text2), true))
-}
-
 func TestLoadDictionary(t *testing.T) {
-
 	var seg, seg1 Segmenter
 
 	ToLower = false
@@ -228,20 +103,22 @@ func TestLoadDictionary(t *testing.T) {
 
 	err := seg.LoadDict("en")
 	tt.Nil(t, err)
+	seg.Load = false
 
 	err = seg.LoadDict("zh")
 	tt.Nil(t, err)
 
 	err = seg1.LoadDict("jp")
 	tt.Nil(t, err)
+	seg1.Load = false
 
-	err = prodSeg.LoadDict()
+	err = seg1.LoadDict()
 	tt.Nil(t, err)
 
 	tt.Expect(t, "世界/n 人口/n ", ToString(prodSeg.Segment(
 		[]byte("世界人口")), false))
 
-	dict := prodSeg.Dictionary()
+	dict := seg1.Dictionary()
 
 	freq, ok := dict.Find([]byte("世界"))
 	tt.Equal(t, 34387, freq)
@@ -255,11 +132,11 @@ func TestLoadDictionary(t *testing.T) {
 	tt.Equal(t, 3, freq)
 	tt.True(t, ok)
 
-	freq, ok = prodSeg.Find("帝国大厦大")
+	freq, ok = seg1.Find("帝国大厦大")
 	tt.Equal(t, 0, freq)
 	tt.False(t, ok)
 
-	val, id, err := prodSeg.Value("帝国")
+	val, id, err := seg1.Value("帝国")
 	tt.Equal(t, 147102, val)
 	tt.Equal(t, 256572, id)
 	tt.Nil(t, err)
@@ -286,6 +163,19 @@ func TestLoadDictionary(t *testing.T) {
 	freq, ok = dict.Find([]byte("西雅图太空针"))
 	tt.Equal(t, 0, freq)
 	tt.False(t, ok)
+}
+
+func TestDictPaths(t *testing.T) {
+	SkipLog = true
+	paths := DictPaths("./dictDir", "zh, jp")
+	tt.Expect(t, "2", len(paths))
+
+	tt.Expect(t, "dictDir/dict/dictionary.txt", paths[0])
+	tt.Expect(t, "dictDir/dict/jp/dict.txt", paths[1])
+
+	paths1 := DictPaths("./dictDir", "zh, jp")
+	tt.Expect(t, "2", len(paths))
+	tt.Equal(t, paths, paths1)
 }
 
 func TestInAlphaNum(t *testing.T) {
@@ -316,54 +206,4 @@ func TestInAlphaNum(t *testing.T) {
 	tx = seg.Cut(text, true)
 	tt.Equal(t, 9, len(tx))
 	tt.Equal(t, "[hello world !  你好 世界 ,  Hello world .]", tx)
-}
-
-var token = Token{
-	text: []Text{
-		[]byte("one"),
-		[]byte("two"),
-	},
-}
-
-func TestTokenEquals(t *testing.T) {
-	tt.True(t, token.Equals("onetwo"))
-}
-
-func TestTokenNotEquals(t *testing.T) {
-	tt.False(t, token.Equals("one-two"))
-}
-
-var strs = []Text{
-	Text("one"),
-	Text("two"),
-	Text("three"),
-	Text("four"),
-}
-
-func TextSliceToString(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		textSliceToString(strs)
-	}
-}
-
-func TextToString(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		textToString(strs)
-	}
-}
-
-func TestBenchmark(t *testing.T) {
-	fmt.Println("textToString: ")
-	fmt.Println(testing.Benchmark(TextToString))
-
-	fmt.Println("textSliceToString: ")
-	fmt.Println(testing.Benchmark(TextSliceToString))
-}
-
-func BenchmarkEquals(t *testing.B) {
-	fn := func() {
-		token.Equals("onetwo")
-	}
-
-	tt.BM(t, fn)
 }
